support more param types in generated client
since 3.10 aiohttp uses a stricter definition of parameters.
(it slightly changed in patch releases but we can ignore that)
It follows the type definitions of yarl which only allows str, int and
float variables.
We use UUIDs and they don’t work with this definition so this commit
adds a fallback that calls str for scalar types that are not defined in
yarl’s SimpleQuery.